requirements.txt
numpy
pint
pygame
setuptools
Shapely
sleipnirgroup-jormungandr
sortedcontainers
pytest
pytest-cov
pymupdf
google-generativeai
reportlab
